+++ app/src/main/java/com/example/android/quakereport/QueryUtils.java
@@ -58,7 +58,7 @@ public final class QueryUtils {
 
                 double magnitude = properties.getDouble("mag");
                 String location = properties.getString("place");
-                int time = properties.getInt("time");
+                long time = properties.getLong("time");
 
                 earthquakes.add(new Earthquake(magnitude, location, time));
             }
